+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Please provide all feedback to me via https://github.com/wsimmonds/crewai-lint/issues or email wjsimmonds@gmail.com.
 
+## 0.102.1
+- Fix the fact extension would fail to fail to run due to missing dependency.
+- Change "No schema available" error on first load to be info and more helpful message (pending fix/retry logic)
+
 ## 0.102.0
 
 - Initial release featuring basic validation of agents.yaml and tasks.yaml